improve extract_i64 impl
